Add files via upload
This notebook is created to show the processing loops to get school locations - I made it to pull everything from the events data frame - so you could pull the lat lng data either in the API loop or right after the data is pulled and the events data frame is being created.  You can put Lat and Lng data directly into the Events data frame, but you probably will still have to have a location data frame that just has Lat and Lng to put into the gmap function.  I made sure the heat map will run with the way the location data was generated.  I did not clean out the conference locations - it did not throw the data off because the conferences had locations associated with them.  An improvement to the application would be to filter out the conference locations by either pulling the home school from the other column data or filtering it out altogether.  CBS would probably want the school data for a conference - this would take a dedicated function to try to guess at the actual school location.  Let me know that this copy will work for you.  Eve